+++ src/base/Transaction.php
@@ -2,7 +2,7 @@
 
 namespace luya\payment\base;
 
-use luya\payment\PaymentProcess;
+use Curl\Curl;
 use yii\web\Controller;
 use yii\base\BaseObject;
 use luya\payment\Pay;
@@ -13,33 +13,47 @@ use luya\payment\PaymentException;
  *
  * Each transaction must implement the Transaction Abstraction class.
  *
+ * @property PayModel $model
+ * @property Controller $context
+ * @property IntegratorInterface $integrator
+ * 
  * @author Basil Suter <basil@nadar.io>
  * @since 1.0.0
  */
 abstract class Transaction extends BaseObject
 {
     /**
-     * Creates the transaction and mostly redirects to the provider afterwards
+     * Creates the transaction and mostly redirects to the provider afterwards.
      */
     abstract public function create();
     
     /**
-     * Return from create into the back
+     * The method which is triggered when coming "back from the provider". In generall this is also success.
+     * 
+     * Commonly call `$this->redirectApplicationSuccess()` now.
      */
     abstract public function back();
     
     /**
-     * Some providers provide a notify link
+     * Some providers provide a notify link. The notify method will be called from the payment provider in the 
+     * background. In generall we want to ensure the payment is closed here.
+     * 
+     * close the paymentsuccessful `$this->closePaymentAsSuccessful()` and try to call the webstore url (maybe 
+     * this triggers an email) `$this-curlApplicationLink($this->getModel()->getApplicationSuccessLink())`.
      */
     abstract public function notify();
     
     /**
-     * An error/failure happend
+     * Redirect back to the application failure/error page
+     * 
+     * `$this->redirectTransactionFail()`
      */
     abstract public function fail();
     
     /**
      * All providers provide an abort/stop link to back into the onlinestore and choose
+     * 
+     * `$this-redirectTransactionAbort()`
      */
     abstract public function abort();
 
@@ -52,6 +66,24 @@ abstract class Transaction extends BaseObject
      * @var string Certain transactions allows you to configure a title for the payment. For example `John Doe's Estore`.
      */
     public $title;
+
+    /**
+     * @var string The error message which is thrown an logged when an eror happens while closing the model.
+     * @since 2.1
+     */
+    public $errorCloseSuccess = "Unable to close the model as successful, maybe its already closed.";
+
+    /**
+     * @var string The error message which is thrown an logged when an eror happens while closing the model.
+     * @since 2.1
+     */
+    public $errorCloseAbort = "Unable to close the model as aborted, maybe its already closed.";
+
+    /**
+     * @var string The error message which is thrown an logged when an eror happens while closing the model.
+     * @since 2.1
+     */
+    public $errorCloseError = "Unable to close the model as errored, maybe its already closed.";
 
     private $_model;
     
@@ -176,11 +208,7 @@ abstract class Transaction extends BaseObject
     {
         $url = $this->getModel()->getApplicationSuccessLink();
         
-        $closable = $this->getIntegrator()->closeModel($this->getModel(), Pay::STATE_SUCCESS);
-
-        if (!$closable) {
-            throw new PaymentException("Unable to close the model, maybe its already closed.");
-        }
+        $this->closePaymentAsSuccessful();
 
         return $this->getContext()->redirect($url);
     }
@@ -193,11 +221,7 @@ abstract class Transaction extends BaseObject
     public function redirectApplicationAbort()
     {
         $url = $this->getModel()->getApplicationAbortLink();
-        $closable = $this->getIntegrator()->closeModel($this->getModel(), Pay::STATE_ABORT);
-
-        if (!$closable) {
-            throw new PaymentException("Unable to close the model, maybe its already closed.");
-        }
+        $this->closePaymentAsAborted();
 
         return $this->getContext()->redirect($url);
     }
@@ -210,12 +234,80 @@ abstract class Transaction extends BaseObject
     public function redirectApplicationError()
     {
         $url = $this->getModel()->getApplicationErrorLink();
+        $this->closePaymentAsErrored();
+
+        return $this->getContext()->redirect($url);
+    }
+
+    /**
+     * Close the current payment model as successful
+     *
+     * @since 2.1
+     */
+    protected function closePaymentAsSuccessful()
+    {
+        $closable = $this->getIntegrator()->closeModel($this->getModel(), Pay::STATE_SUCCESS);
+
+        if (!$closable) {
+            $this->getIntegrator()->addTrace($this->getModel(), __METHOD__, $this->errorCloseSuccess);
+            throw new PaymentException($this->errorCloseSuccess);
+        }
+    }
+
+    /**
+     * Close the current payment model as aborted
+     *
+     * @since 2.1
+     */
+    protected function closePaymentAsAborted()
+    {
+        $closable = $this->getIntegrator()->closeModel($this->getModel(), Pay::STATE_ABORT);
+
+        if (!$closable) {
+            $this->getIntegrator()->addTrace($this->getModel(), __METHOD__, $this->errorCloseAbort);
+            throw new PaymentException($this->errorCloseAbort);
+        }
+    }
+
+    /**
+     * Close the current payment model as errored
+     *
+     * @since 2.1
+     */
+    protected function closePaymentAsErrored()
+    {
         $closable = $this->getIntegrator()->closeModel($this->getModel(), Pay::STATE_ERROR);
 
         if (!$closable) {
-            throw new PaymentException("Unable to close the model, maybe its already closed.");
+            $this->getIntegrator()->addTrace($this->getModel(), __METHOD__, $this->errorCloseError);
+            throw new PaymentException($this->errorCloseError);
+        }
+    }
+
+    /**
+     * CURL a given url in order to ensure estore "success" page is called.
+     *
+     * @param string $link
+     * @return boolean Whether success status code is returned or not.
+     * @since 2.1
+     */
+    protected function curlApplicationLink($link)
+    {
+        $curl = new Curl();
+        $curl->get($link);
+
+        $this->getIntegrator()->addTrace($this->getModel(), __METHOD__, $link . ' | Response Status Code: ' . $curl->getHttpStatus());
+
+        if ($curl->curl_error) {
+            $this->getIntegrator()->addTrace($this->getModel(), __METHOD__, $link . ' | CURL Response: ' . $curl->curl_error_message);
         }
 
-        return $this->getContext()->redirect($url);
+        $success = $curl->isSuccess();
+
+        if (!$success) {
+            $this->getIntegrator()->addTrace($this->getModel(), __METHOD__, $link . ' | Request Response: ' . $curl->error_message);
+        }
+
+        return $success;
     }
 }